feat(term size): add max-font-size mode
The max-font-size mode behaviors like the fixed-font-size mode
in that a basic font size can be specified by the user.

However, when the window is too narrow or too short,
fixed-font-size just lets the terminal screen overflow the window,
while max-font-size shrinks the font to keep the screen inside the window.

This feature should benefit users
who use PttChrome on a relatively small screen (e.g., mobile/tablet users)
and want to maximize the use of the limited space of their screen for reading
without manually tweaking the terminal size in the fixed-term-size mode.

* src/components/ContextMenu/PrefModal.js PrefModal
    * Add "max-font-size" mode as a new termSizeMode
    * Add "options_maxFontSize" to the "termSizeMode" select list

* src/js/en_US_messages.js & src/js/zh_TW_messages.js
    * Add translated messages for "options_maxFontSize" options

* src/js/pttchrome.js App.prototype.onValuesPrefChange()
    * Add fontsize-handling logic to the new termSizeMode case 'max-font-size'

* src/js/term_view.js
    * Add TermView.prototype.calcFontSizeFromTerm()
      to obtain the maximum allowed font size for a given terminal size