feat(TwoColorWord.js): implement blinking
TwoColorWord achieves the two-color-word effect
by adding a half-word overlay to a full-width character ("word").

Therefore, the blinking effect is implemented with the 3 cases:
* Only the tail blinks: Overlay the head and blink only the underneath word
* Only the head blinks: Overlay the tail and blink only the underneath word
* Both blink: Overlay the head and blink both the underneath word and the overlay

For overlaying the tail,
a negative text-indent is used to truncate off the left half of the overlay.

The background color is handled by the .b* CSS classes
but was also affected by the .qq* classes,
which caused another point of complexity
for implementing the TwoColorWord blinking.

Therefore,
only the .qq class (which does not affect the background color) is kept,
while the other .qq* classes are removed.

* src/components/Row/WordSegmentBuilder/ColorSpan.js
    * Use CSS class .qq for the blinking effect
* src/components/Row/WordSegmentBuilder/TwoColorWord.js
    * Add implementation description
    * Implement blinking
* src/css/color.css
    * Use class .qq for the blinking effect solely and remove other .qq* classes