chore(package.json): bump version number to 2.11.0-ccns.2021
The current version should be > 1.5.3 according to the git tags,
but the version number was never updated in package.json since 1.2.0.

Git tags before this commit were added using a semver-like versioning scheme.
See https://github.com/ccns/PttChrome/tags.

"-ccns.2021" denotes that this version is on the branch "main-ccns.2021",
which was rebased onto the branch `dev` in 2021 and is maintained by CCNS.